+++ host-tests/src/android/util/Log.java
@@ -1,0 +1,10 @@
+package android.util;
+
+public final class Log {
+    private Log() {
+    }
+
+    public static int e(String tag, String message) {
+        return 0;
+    }
+}